FIX: Share page genera copias en directorio temporal en vez de permanentes. ADD: Inyección de Context. ADD: Peticion POST para procesar imagen. FIX: Corregido alineamiento del Canvas a la imagen. MODIFY: Creado directorio Mappers

+++ app/src/main/java/com/mastermovilesua/persistencia/tfm_detectorpentagramas/data/network/model/BoxModel.kt
@@ -4,8 +4,8 @@ import com.google.gson.annotations.SerializedName
 import com.mastermovilesua.persistencia.tfm_detectorpentagramas.data.network.contracts.PageApiContract
 
 data class BoxModel (
-    @SerializedName(PageApiContract.BOX_FIELD_X) val x: Int,
-    @SerializedName(PageApiContract.BOX_FIELD_Y) val y: Int,
-    @SerializedName(PageApiContract.BOX_FIELD_WIDTH) val width: Int,
-    @SerializedName(PageApiContract.BOX_FIELD_HEIGHT) val height: Int
+    @SerializedName(PageApiContract.BOX_FIELD_X) val x: Float,
+    @SerializedName(PageApiContract.BOX_FIELD_Y) val y: Float,
+    @SerializedName(PageApiContract.BOX_FIELD_WIDTH) val width: Float,
+    @SerializedName(PageApiContract.BOX_FIELD_HEIGHT) val height: Float
 )
